+++ src/tasks/tslint/__tests__/global-configuration/gulpfile.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const gulp = require('gulp');
+const task = require('../../index');
+
+task.dependencies = {};
+task.register(gulp, 'tslint', true, {
+  sourceFiles: 'index-invalid.ts'
+});