+++ src/tasks/tslint/__tests__/valid-project/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Valid
+ */
+export function index(): string {
+  return 'valid';
+}